rename "MQClient" -> "mqclient"
