+++ integration_tests/common_backend_interface_tests.py
@@ -9,7 +9,7 @@ import logging
 from typing import List, Optional
 
 # local imports
-from MQClient.backend_interface import Backend, Message
+from mqclient.backend_interface import Backend, Message
 
 from .utils import DATA_LIST, _log_recv, _log_send
