+++ setup.cfg
@@ -29,10 +29,10 @@ directory = htmlcov
 ignore=E226,E261,E302,E305,E501,W503,W504
 
 [tool:pytest]
-testpaths = MQClient tests
+testpaths = mqclient tests
 
 [semantic_release]
-version_variable = MQClient/__init__.py:__version__
+version_variable = mqclient/__init__.py:__version__
 upload_to_pypi = False
 patch_without_tag = True
 major_on_zero = False
