+++ tests/test_init.py
@@ -5,5 +5,5 @@ import importlib
 
 def test_import() -> None:
     """Test module imports."""
-    m = importlib.import_module("MQClient")
+    m = importlib.import_module("mqclient")
     assert hasattr(m, "Queue")
